+++ HISTORY.md
@@ -122,10 +122,11 @@
     * Catalan, thanks, Joan!
     * German
 
-### 2026-04-28
+### 2026-04-29
 * Catalan translation update, thank you, Joan!
 * objects: fixed applying graphics effects to sprites
 * fixed initialization of the "Shapes" library, thanks, @DarDoro, for the report!
+* objects: added basic support for writing in "overdraw" mode on a costume-less sprite
 
 ### 2026-04-28
 * store: refactored loading local custom blocks with custom dropdowns in variadic slots